🧹 Code Health: Add explicit unsigned int cast in user-vcsm
🎯 What: Added an explicit (unsigned int) cast to elem->dma_addr in vcsm_vc_addr_from_hdl.
💡 Why: To fulfill a code health TODO and ensure explicit type conversion when returning a 32-bit DMA address (uint32_t) from a function returning unsigned int.
✅ Verification: Compiled a standalone test stub and recompiled the full native build with strict flags (-Werror=format). No new warnings or errors were introduced.
✨ Result: Improved code clarity and safety by explicitly acknowledging the conversion instead of relying on implicit conversions, adhering to the codebase style.

+++ host_applications/linux/libs/sm/user-vcsm.c
@@ -1061,7 +1061,7 @@ unsigned int vcsm_vc_addr_from_hdl( unsigned int handle )
                       __func__, handle);
          return 0;
       }
-      return elem->dma_addr;
+      return (unsigned int)elem->dma_addr;
    }
    else
    {
